Performance is all that matters, Rust is a systems programming language.

This commit implements the following optimisations:
  1. Avoid using the `format!` machinery (via `println!` in a tight loop).
  2. Avoid allocating a `Vec` of arguments when we just need the second one.
  3. Avoid locking `stdout` over and over again in a tight loop, by hoisting
    `stdout.lock()` outside the loop.
  4. Use buffered output.

Points 1, 3 and 4 improve overall performance, while point 2 improves start up speed. **Please review carefully.**

This commit increases output throughput as measured by (*short*)

    target/release/yes | pv > /dev/null

and (*long*)

    target/release/yes aaaaaaaaaaaaaaaa | pv > /dev/null

after 20s from:

| Implementation | Throughput (*short*) | Throughput (*long*) |
| -------------------- | ------ | ------------- |
| Old |  3.8 MB/s | 30.2 MB/s |
| New | **4.5 GB/s**| **4.5 GB/s**|
| GNU coreutils 8.21 | 140MB/s | 769MB/s|

This commit also improves error handling by breaking out of the loop on an I/O error.

This PR is also ridiculous.

**EDIT**

I added two new commits with the following additional optimisations:
    1. Do not allocate in the `"y\n"` case by using `Cow` (suggested by @eddyb and @withoutboats)
    2. Use `memcpy`-doubling to write progressively bigger buffers. Starts with a single copy of the output, then two, three etc. up until 64KB (suggested by @Veedrac).

Point (2) in particular increased throughput by an order of magnitude and it has the added advantage that time to first byte is really low compared to a fully-buffered approach.


Thanks everyone for contributing. It is very important to get this sort of thing right or people will never migrate off of C++.


**EDIT the second**

Now we also support non-unicode arguments on Unix!