@nathanross 2016-08-06
hey @cgati and @cristicbz , would you be willing to license to code in an MIT-compatible way? This is definitely higher throughput (having just tested) than the current version in [uutils/coreutils](https://github.com/uutils/coreutils/blob/a2944e06db93c70f20556fdd5039abfabce78be6/src/yes/yes.rs)