@Mark-Simulacrum 2016-08-06
I get speeds of 10.4 GB/s with `yes` and 10.2 GB/s with `yes yyyyyyyyyyyyyyyyyyyy` with the following code. GNU coreutils v8.25 has exactly the same speed with the `yes` and 10.4 GB/s with the long invocation. I'm not sure why it wins with the long invocation.

With the current code in this PR, I get ~790 MB/s with the short invocation and ~1.1 GB/s with the long invocation. The problem here seems to lie in that BufWriter's `write` implementation will be fastest when writing the bytes immediately; I've removed buffering entirely to get the speeds above. It's possible the startup time could be further optimized through reducing allocations.

```rust
use std::env;
use std::io::{self, Write};
use std::process;

const PER_WRITE: usize = 8 * 1024;

fn main() {
    {
        // Move all locals into a scope, so that `process::exit()` is called after all destructors.
        let output = env::args().nth(1)
            .map(|x| x + "\n")
            .unwrap_or("y\n".to_owned());
        let stdout = io::stdout();
        let mut locked = stdout.lock();

        let len = PER_WRITE / output.len();
        let buffer = std::iter::repeat(output).take(len).collect::<String>();

        loop {
            if locked.write(buffer.as_bytes()).is_err() {
                break;
            }
        }
    }
    process::exit(1);
}
```