@Stebalien 2016-08-06
FYI, constructing the buffer that way will reallocate the string multiple times (collect has no way of knowing how long each of the `output` strings is). Instead, I recommend:

```rust
let mut buffer = String::with_capacity(PER_WRITE);
buffer.extend(std::iter::repeat(output).take(len));
```

(although the syscalls will dominate anyways so this is probably a useless optimization...)