@Mark-Simulacrum 2016-08-06
That may have helped with startup time, difficult to check...
However, I did just check another thing: `while ....is_ok()` instead of the loop/break in the code above is slower.