@serprex 2016-08-06
If the string is smaller than 4kb then it should be repeated to that size. Then print with http://mainisusuallyafunction.blogspot.ca/2014/09/raw-system-calls-for-rust.html (syscalls crate now supports more than x86_64 linux)