@eddyb 2016-08-06
@serpex Unbuffered stdout calls `write` which ends up in the syscall. There should be no overhead other than the calls themselves.