@cgati 2016-08-06
@nathanross absolutely. I'm currently out of town without my laptop for my sister's graduation, but as soon as I'm home this evening I'd be happy to.