@Veedrac 2016-08-06
That code fails if the input string is longer than the buffer. You should do

    let len = PER_WRITE / output.len() + 1;

to avoid that. Further, all of these break if `locked.write` doesn't write all of the bytes - they should be using `write_all`. That can have a performance overhead (I dunno, I never checked), so really you want:

    let len = std::cmp::max(PER_WRITE, output.len());
    let buffer: Vec<u8> = output.bytes().cycle().take(len).collect();

    let mut offset = 0;
    while let Ok(bytes) = locked.write(&buffer[offset..]) {
        offset += bytes;
        offset %= output.len();
    }

You probably also want to use `args_os`, since `yes` doesn't require unicode. I don't know a pleasant platform-independant way to turn an `OsStr` to a `&[u8]`, so you'll need platform-specific code. It just so turns out that `yes` actually takes multiple arguments, too, separating them with spaces. And a `--help` argument, and a `--version` argument... we should support those. And the `yes -- foo` argument syntax. You know, I'm really starting to doubt my sanity. But that's fine... let's let's optimize the startup time too. that would be a Good Idea. How about we dispatch off of the `env::args` and IF it's length one then IF the argument is longer than `PER_WRITE` we just convert it to a `Vec<u8>` directly (except on Windo`ws) b`ecause that avoids generating the buffer but we should Also specialize on zer0 length because we can generate that with specialized code `buffer = vec![b'\n'; PER_WRITE]; for i in 0..PER_WRITE / 2 { buffer[i * 2] = b'y'; }` but then for other lengths we can use m`emcpy `-based doubling like while i < buffer.len() { let (lhs, mut rhs) = buffer.split_at_mut(i); let _ = rhs.write(lhs); i *= 2; } then it's fast.