@cristicbz 2016-08-06
Excellent, thanks a lot @eddyb, @withoutboats, @Stebalien and @Veedrac , I implemented your suggestions, for a grand total of 4.5 GB/s throughput! See 08116fb.