@cristicbz 2016-08-06
Hey @nathanross , despite the extreme innovation displayed in this PR, I will generously dual licence my contributions as MIT/Apache-2.0. Though I imagine that for uutils, you want more features like `--version` etc, right?