@nathanross 2016-08-06
@cristicbz @cgati great! Thank you.

@cristicbz yes, that the uutils ```yes``` will support that flag; I will build the commit next weekend if no one else has by then.