@cristicbz 2016-08-08
I implemented all your suggestions, including `OsString` to support non-unicode strings (though on non-unix, I still roundtrip through UTF-8; we can leave this for another PR). OS specific support is encapsulated in the `platform` module.

Also fixed the edge cases around the buffer size vs output size. 

cc @cgati @llogiq @Stebalien @Stebalien @eddyb @Veedrac 

Come on, let's have one last look, pull together and land this momentous PR.