@cgati 2016-08-08
LGTM. I'll go ahead and merge this in. Thank you everyone for your suggestions and all the work done here!