@mmstick 2016-08-09
The GNU `yes` command actually accepts multiple arguments, instead of just the first argument.