@cgati 2016-08-09
@mmstick I originally wrote this to mimic the BSD `yes` as found on macOS (which is non-obvious, as I never stated that before.) That could probably be filed as an issue, however.