@aij 2016-08-09
@mmstick I was wondering whether that was standard or a GNU extension.

GNU yes also tries to parse options, which I believe is nonstandard. Eg:

    $ yes --verbose
    yes: invalid option -- '-'
    Try 'yes --help' for more information.