@xfix 2016-08-28
`yes` is not POSIX, so it's hard to say what are extensions.